Update server to use request params
According to specification HTTP GET request
can have a body and the first implementation
relied on this fact. However, it turns out
that many clients do not support this idea.

As a more robust and universal solution, I
used params-based approach, so the API
can be consumed by any client, including
wen browsers.